refactor(deps): replace vendored ramose with package dependency
