+++ pyproject.toml
@@ -41,7 +41,7 @@ dependencies = [
     "python-lsp-jsonrpc>=1.1.2",
     "python-lsp-server>=1.12.2",
     "pytz>=2026.1.post1",
-    "ramose>=1.0.8",
+    "ramose>=2.0.0",
     "rdflib>=6.1.1",
     "rdflib-jsonld>=0.6.2",
     "redis>=7.0.1",
@@ -75,7 +75,6 @@ pythonpath = "."
 
 [tool.coverage.run]
 source = ["src"]
-omit = ["src/ramose.py"]
 
 [tool.hatch.build.targets.wheel]
 packages = ["src"]
